Get offscreen canvas data back to main thread (WIP)
Stopped rendering to a RenderTarget, because I couldn't figure out how
to do anything useful with it. This is really more evidence that this
particular approach is broken.

Worked out that I can convert the OffscreenCanvas to a BLOB, then
convert that BLOB to a ByteArray, and *then* transfer the ByteArray
back to the main thread.

I'm not sure what options I have available to deal with it from
there.

I'm kinda/sorta setting things up to build a three.js Texture from
that raw data, but I think I'd need to implement a custom Loader.

I know this approach is not viable long-term. It probably isn't viable
short-term. It doesn't seem worth the effort it would take to verify
that.

For #roadblocks