Switch to transferring ImageBitmaps
This is exactly what they're for.

My spinning cube is now a plain gray box rather than the "Loading
texture" placeholder I've been staring at for so long.

It's a far cry from "good," but it's also a huge step forward and
makes me feel much better about the prospects of this approach.