Switched from devcards to NuBank workspaces (WIP)
For #roadblocks because I want to preview what I'm trying to draw.

I think this approach is probably a dead end. Rendering react is
conceptually similar to running a block of webgl code, but I think
this is probably a square peg/round hole situation.

I really should spend time auditing the Workspace code base to truly
convince myself of that.

Allowed npm to fix a couple of security issues with library versions
while I had my REPL down. Just to add to the confusion.