Get preview scenes to animate
And experiment with background color clearing. This is obnoxious
enough at low frame rates that it very well trigger a seizure for
anyone who's prone to that sort of thing.

But at least I have a vague notion about why the scene backgrounds and
renderer clear color interact the way they do.

For #roadblocks