Add a runner and animate it along that new track
The way the "physics" is set up, I needed to add the previous timestamp
to the animation frame callback so I can hand it a time delta.

The preview card looks fairly reasonable in firefox.

It looks more like a snarl of Christmas lights in chromium.

Eliminated the "red" scene I started with as a proof of concept. I'm
not quite comfortable enough to also get rid of the "blue" scene, but
having both was just silly.