Refactor preview track creation pieces
Move them into their own function so I can create another preview,
from the Runner's point of view.

While I was at it, I changed the way the track curve gets
created. From the caller's viewpoint, you pass in a seq of numeric
tuples to specify the control points.

That way, it doesn't have to a) create the spline and then b) use the
spline to create the track.

It still needs to be aware of the spline, so it can follow along. But
at least it's a little less tightly coupled to the implementation.