Rotate runner & camera along track
Added a CameraHelper to the 3rd-person view to help figure out what
the camera is doing.

It still isn't really following the way I want, but this is a huge
improvement over the previous commit.

Switched to a thicker dashe line for the track, though it isn't doing
what I expected.

Oh, and got rid of my preliminary blue scene as a complete waste.

For #roadblocks