Try to get keyboard input from preview pane (WIP)
To trigger a jump event

This is basically a pointless commit that doesn't amount to much of
anything. But I've been wrestling with the basic idea for days now,
and I need to switch to a different branch to add an event queue to
send the signal that this will [hypothetically] trigger.

For #roadblocks